requirements.txt
Django==3.2.5
djangorestframework==3.12.4
djangorestframework_simplejwt == 5.5.0
